unternehmensgründung/backend/private_functions/requirements.txt
hdwallet
bit
mysql-connector